# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"Feira Hub API","description":"Backend para catálogo de produtos e provedores consumido por frontend React.","version":"1.0.0"},"paths":{"/feed/provedores":{"get":{"tags":["Feed"],"summary":"Listar Todos Provedores","description":"READ - Todos os provedores cadastrados.","operationId":"listar_todos_provedores_feed_provedores_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ProvedorResponse"},"type":"array","title":"Response Listar Todos Provedores Feed Provedores Get"}}}}}}},"/feed/carrossel":{"get":{"tags":["Feed"],"summary":"Listar Provedores Carrossel","description":"READ - Provedores que possuem foto_banner preenchida para o carrossel.","operationId":"listar_provedores_carrossel_feed_carrossel_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ProvedorResponse"},"type":"array","title":"Response Listar Provedores Carrossel Feed Carrossel Get"}}}}}}},"/feed/itens/mais-vendidos":{"get":{"tags":["Feed"],"summary":"Listar Itens Mais Vendidos","description":"READ - Todos os itens ordenados pelos mais vendidos (ORDER BY vendidos DESC).","operationId":"listar_itens_mais_vendidos_feed_itens_mais_vendidos_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ItemResponse"},"type":"array","title":"Response Listar Itens Mais Vendidos Feed Itens Mais Vendidos Get"}}}}}}},"/feed/vendedor/{nome_provedor}/itens":{"get":{"tags":["Feed"],"summary":"Listar Itens Por Vendedor","description":"READ - Todos os itens pertencentes a um provedor/vendedor específico.","operationId":"listar_itens_por_vendedor_feed_vendedor__nome_provedor__itens_get","parameters":[{"name":"nome_provedor","in":"path","required":true,"schema":{"type":"string","title":"Nome Provedor"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ItemResponse"},"title":"Response Listar Itens Por Vendedor Feed Vendedor  Nome Provedor  Itens Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/feed/cidade/{nome_cidade}/itens":{"get":{"tags":["Feed"],"summary":"Listar Itens Por Cidade","description":"READ - Itens dos provedores de uma cidade, ordenados pelos mais vendidos.","operationId":"listar_itens_por_cidade_feed_cidade__nome_cidade__itens_get","parameters":[{"name":"nome_cidade","in":"path","required":true,"schema":{"type":"string","title":"Nome Cidade"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ItemResponse"},"title":"Response Listar Itens Por Cidade Feed Cidade  Nome Cidade  Itens Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/busca/itens":{"get":{"tags":["Busca"],"summary":"Buscar Itens","description":"Busca itens por nome (WHERE nome_item LIKE %q%).","operationId":"buscar_itens_busca_itens_get","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string","description":"Termo para buscar no nome do item","title":"Q"},"description":"Termo para buscar no nome do item"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ItemResponse"},"title":"Response Buscar Itens Busca Itens Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/busca/vendedores":{"get":{"tags":["Busca"],"summary":"Buscar Vendedores","description":"Busca provedores onde categoria = 'vendedor' e nome LIKE %q%.","operationId":"buscar_vendedores_busca_vendedores_get","parameters":[{"name":"q","in":"query","required":false,"schema":{"type":"string","description":"Termo para buscar no nome do vendedor","default":"","title":"Q"},"description":"Termo para buscar no nome do vendedor"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProvedorResponse"},"title":"Response Buscar Vendedores Busca Vendedores Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/busca/servicos":{"get":{"tags":["Busca"],"summary":"Buscar Servicos","description":"Busca provedores onde categoria = 'prestador de serviço' e nome LIKE %q%.","operationId":"buscar_servicos_busca_servicos_get","parameters":[{"name":"q","in":"query","required":false,"schema":{"type":"string","description":"Termo para buscar no nome do prestador de serviço","default":"","title":"Q"},"description":"Termo para buscar no nome do prestador de serviço"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProvedorResponse"},"title":"Response Buscar Servicos Busca Servicos Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/itens/":{"post":{"tags":["Painel do Provedor - Itens"],"summary":"Criar novo item para um provedor","description":"CREATE - Cadastra um item vinculado a um provedor.","operationId":"criar_item_itens__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Criar Item Itens  Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/itens/provedor/{nome_provedor}":{"get":{"tags":["Painel do Provedor - Itens"],"summary":"Listar itens de um provedor específico","description":"READ (Listar) - Retorna todos os itens cadastrados pelo provedor.","operationId":"listar_itens_provedor_itens_provedor__nome_provedor__get","parameters":[{"name":"nome_provedor","in":"path","required":true,"schema":{"type":"string","title":"Nome Provedor"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ItemResponse"},"title":"Response Listar Itens Provedor Itens Provedor  Nome Provedor  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/itens/{nome_item}":{"get":{"tags":["Painel do Provedor - Itens"],"summary":"Obter detalhes de um item","description":"READ - Retorna os dados de um item específico pela chave primária.","operationId":"obter_item_itens__nome_item__get","parameters":[{"name":"nome_item","in":"path","required":true,"schema":{"type":"string","title":"Nome Item"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Painel do Provedor - Itens"],"summary":"Atualizar dados de um item","description":"UPDATE - Atualiza foto, preços e estoque do item.","operationId":"atualizar_item_itens__nome_item__put","parameters":[{"name":"nome_item","in":"path","required":true,"schema":{"type":"string","title":"Nome Item"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Atualizar Item Itens  Nome Item  Put"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Painel do Provedor - Itens"],"summary":"Deletar um item","description":"DELETE - Remove um item do catálogo.","operationId":"deletar_item_itens__nome_item__delete","parameters":[{"name":"nome_item","in":"path","required":true,"schema":{"type":"string","title":"Nome Item"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Deletar Item Itens  Nome Item  Delete"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/provedores/":{"get":{"tags":["Painel do Administrador - Provedores"],"summary":"Listar todos os provedores/vendedores","description":"READ (Listar) - Retorna todos os provedores cadastrados.","operationId":"listar_todos_provedores_provedores__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ProvedorResponse"},"type":"array","title":"Response Listar Todos Provedores Provedores  Get"}}}}}},"post":{"tags":["Painel do Administrador - Provedores"],"summary":"Cadastrar novo provedor/vendedor","description":"CREATE - Cadastra um novo provedor ou prestador de serviço.","operationId":"criar_provedor_provedores__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProvedorCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Criar Provedor Provedores  Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/provedores/{nome_provedor}":{"get":{"tags":["Painel do Administrador - Provedores"],"summary":"Obter detalhes de um provedor","description":"READ - Retorna os dados de um provedor específico pela PK.","operationId":"obter_provedor_provedores__nome_provedor__get","parameters":[{"name":"nome_provedor","in":"path","required":true,"schema":{"type":"string","title":"Nome Provedor"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProvedorResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Painel do Administrador - Provedores"],"summary":"Atualizar dados de um provedor","description":"UPDATE - Atualiza as informações cadastrais do provedor.","operationId":"atualizar_provedor_provedores__nome_provedor__put","parameters":[{"name":"nome_provedor","in":"path","required":true,"schema":{"type":"string","title":"Nome Provedor"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProvedorUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Atualizar Provedor Provedores  Nome Provedor  Put"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Painel do Administrador - Provedores"],"summary":"Deletar um provedor","description":"DELETE - Remove um provedor do sistema.","operationId":"deletar_provedor_provedores__nome_provedor__delete","parameters":[{"name":"nome_provedor","in":"path","required":true,"schema":{"type":"string","title":"Nome Provedor"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Deletar Provedor Provedores  Nome Provedor  Delete"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/provedores/{nome_provedor}/faturamento":{"get":{"tags":["Painel do Administrador - Provedores"],"summary":"Consultar total faturado por vendedor","description":"Calcula a soma de vendas em dinheiro registradas para o vendedor:\nSELECT SUM(valor_total) FROM vendas WHERE provedor = ?","operationId":"obter_faturamento_vendedor_provedores__nome_provedor__faturamento_get","parameters":[{"name":"nome_provedor","in":"path","required":true,"schema":{"type":"string","title":"Nome Provedor"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Obter Faturamento Vendedor Provedores  Nome Provedor  Faturamento Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/carrinho/sync":{"post":{"tags":["Carrinho / Checkout WhatsApp"],"summary":"Atualiza vendas e estoque a partir do carrinho do React","description":"Recebe os itens selecionados no React no momento do redirecionamento\npara o WhatsApp e aplica o incremento no campo 'vendidos' e decremento no 'estoque'.","operationId":"sincronizar_carrinho_carrinho_sync_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CarrinhoSyncRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Sincronizar Carrinho Carrinho Sync Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"tags":["Health Check"],"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"CarrinhoSyncRequest":{"properties":{"itens":{"items":{"$ref":"#/components/schemas/ItemCarrinhoUpdate"},"type":"array","title":"Itens"}},"type":"object","required":["itens"],"title":"CarrinhoSyncRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ItemCarrinhoUpdate":{"properties":{"nome_item":{"type":"string","title":"Nome Item"},"quantidade":{"type":"integer","exclusiveMinimum":0.0,"title":"Quantidade"}},"type":"object","required":["nome_item","quantidade"],"title":"ItemCarrinhoUpdate"},"ItemCreate":{"properties":{"nome_item":{"type":"string","title":"Nome Item"},"provedor":{"type":"string","title":"Provedor"},"foto":{"type":"string","title":"Foto","default":""},"preco_varejo":{"type":"number","minimum":0.0,"title":"Preco Varejo"},"preco_atacado":{"type":"number","minimum":0.0,"title":"Preco Atacado"},"estoque":{"type":"integer","minimum":0.0,"title":"Estoque","default":0},"vendidos":{"type":"integer","minimum":0.0,"title":"Vendidos","default":0}},"type":"object","required":["nome_item","provedor","preco_varejo","preco_atacado"],"title":"ItemCreate"},"ItemResponse":{"properties":{"nome_item":{"type":"string","title":"Nome Item"},"provedor":{"type":"string","title":"Provedor"},"foto":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Foto"},"preco_varejo":{"type":"number","title":"Preco Varejo"},"preco_atacado":{"type":"number","title":"Preco Atacado"},"estoque":{"type":"integer","title":"Estoque"},"vendidos":{"type":"integer","title":"Vendidos"}},"type":"object","required":["nome_item","provedor","preco_varejo","preco_atacado","estoque","vendidos"],"title":"ItemResponse"},"ItemUpdate":{"properties":{"foto":{"type":"string","title":"Foto","default":""},"preco_varejo":{"type":"number","minimum":0.0,"title":"Preco Varejo"},"preco_atacado":{"type":"number","minimum":0.0,"title":"Preco Atacado"},"estoque":{"type":"integer","minimum":0.0,"title":"Estoque","default":0}},"type":"object","required":["preco_varejo","preco_atacado"],"title":"ItemUpdate"},"ProvedorCreate":{"properties":{"nome":{"type":"string","title":"Nome"},"descricao":{"type":"string","title":"Descricao","default":""},"foto_banner":{"type":"string","title":"Foto Banner","default":""},"telefone":{"type":"string","title":"Telefone"},"local":{"type":"string","title":"Local"},"categoria":{"type":"string","enum":["vendedor","prestador de serviço"],"title":"Categoria"},"cidade":{"type":"string","title":"Cidade"}},"type":"object","required":["nome","telefone","local","categoria","cidade"],"title":"ProvedorCreate"},"ProvedorResponse":{"properties":{"nome":{"type":"string","title":"Nome"},"descricao":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Descricao"},"foto_banner":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Foto Banner"},"telefone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Telefone"},"local":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Local"},"categoria":{"type":"string","title":"Categoria"},"cidade":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cidade"}},"type":"object","required":["nome","categoria"],"title":"ProvedorResponse"},"ProvedorUpdate":{"properties":{"descricao":{"type":"string","title":"Descricao","default":""},"foto_banner":{"type":"string","title":"Foto Banner","default":""},"telefone":{"type":"string","title":"Telefone"},"local":{"type":"string","title":"Local"},"categoria":{"type":"string","enum":["vendedor","prestador de serviço"],"title":"Categoria"},"cidade":{"type":"string","title":"Cidade"}},"type":"object","required":["telefone","local","categoria","cidade"],"title":"ProvedorUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}